Move Challenge and ChallengeScalar into the transcript module
