+++ src/transcript.rs
@@ -3,6 +3,7 @@
 
 use ff::Field;
 use std::marker::PhantomData;
+use std::ops::Deref;
 
 use crate::arithmetic::{CurveAffine, FieldExt};
 
@@ -120,5 +121,73 @@ impl<C: CurveAffine, HBase: Hasher<C::Base>, HScalar: Hasher<C::Scalar>>
     pub fn squeeze(&mut self) -> C::Base {
         self.conditional_scalar_squeeze();
         self.base_hasher.squeeze()
+    }
+}
+
+/// This is a 128-bit verifier challenge.
+#[derive(Copy, Clone, Debug)]
+pub struct Challenge(pub(crate) u128);
+
+impl Challenge {
+    /// Obtains a new challenge from the transcript.
+    pub fn get<C, HBase, HScalar>(transcript: &mut Transcript<C, HBase, HScalar>) -> Challenge
+    where
+        C: CurveAffine,
+        HBase: Hasher<C::Base>,
+        HScalar: Hasher<C::Scalar>,
+    {
+        Challenge(transcript.squeeze().get_lower_128())
+    }
+}
+
+/// The scalar representation of a verifier challenge.
+///
+/// The `T` type can be used to scope the challenge to a specific context, or set to `()`
+/// if no context is required.
+#[derive(Copy, Clone, Debug)]
+pub struct ChallengeScalar<F: FieldExt, T> {
+    inner: F,
+    _marker: PhantomData<T>,
+}
+
+impl<F: FieldExt, T> From<Challenge> for ChallengeScalar<F, T> {
+    /// This algorithm applies the mapping of Algorithm 1 from the
+    /// [Halo](https://eprint.iacr.org/2019/1021) paper.
+    fn from(challenge: Challenge) -> Self {
+        let mut acc = (F::ZETA + F::one()).double();
+
+        for i in (0..64).rev() {
+            let should_negate = ((challenge.0 >> ((i << 1) + 1)) & 1) == 1;
+            let should_endo = ((challenge.0 >> (i << 1)) & 1) == 1;
+
+            let q = if should_negate { -F::one() } else { F::one() };
+            let q = if should_endo { q * F::ZETA } else { q };
+            acc = acc + q + acc;
+        }
+
+        ChallengeScalar {
+            inner: acc,
+            _marker: PhantomData::default(),
+        }
+    }
+}
+
+impl<F: FieldExt, T> ChallengeScalar<F, T> {
+    /// Obtains a new challenge from the transcript.
+    pub fn get<C, HBase, HScalar>(transcript: &mut Transcript<C, HBase, HScalar>) -> Self
+    where
+        C: CurveAffine,
+        HBase: Hasher<C::Base>,
+        HScalar: Hasher<C::Scalar>,
+    {
+        Challenge::get(transcript).into()
+    }
+}
+
+impl<F: FieldExt, T> Deref for ChallengeScalar<F, T> {
+    type Target = F;
+
+    fn deref(&self) -> &F {
+        &self.inner
     }
 }
